separate input fns and modeling into separate files
