+++ training/siamese_bert.py
@@ -8,119 +8,11 @@ import pickle
 import pandas as pd
 import numpy as np
 import itertools
+from training.input_fns import input_fn_builder, input_fn_builder_tfrecords
+from training.siamese_modeling import create_model
 from training.ftm_processor import FtmProcessor
 from time import time as tt
 from tensorflow.python import debug as tf_debug
-
-
-def input_fn_builder(features, seq_length, is_training, drop_remainder,
-                     labels=None):
-    """Creates an `input_fn` closure to be passed to TPUEstimator."""
-    assert labels is None or len(features) == len(labels), '''`features` and `labels` should be the same length'''
-
-    all_l_input_ids = []
-    all_r_input_ids = []
-    all_l_input_masks = []
-    all_r_input_masks = []
-
-    for feature in features:
-        all_l_input_ids.append(feature.l_input_ids)
-        all_r_input_ids.append(feature.r_input_ids)
-        all_l_input_masks.append(feature.l_input_mask)
-        all_r_input_masks.append(feature.r_input_mask)
-
-    def input_fn(params):
-        """The actual input function."""
-        batch_size = params["batch_size"]
-
-        num_examples = len(features)
-
-        # This is for demo purposes and does NOT scale to large data sets. We do
-        # not use Dataset.from_generator() because that uses tf.py_func which is
-        # not TPU compatible. The right way to load data is with TFRecordReader.
-        feats_dict = {
-            "l_input_ids":
-                tf.constant(
-                    all_l_input_ids, shape=[num_examples, seq_length],
-                    dtype=tf.int32,
-                    name='l_input_ids'),
-            "r_input_ids":
-                tf.constant(
-                    all_r_input_ids, shape=[num_examples, seq_length],
-                    dtype=tf.int32,
-                    name='r_input_ids'),
-            "l_input_mask":
-                tf.constant(
-                    all_l_input_masks,
-                    shape=[num_examples, seq_length],
-                    dtype=tf.int32,
-                    name='l_input_mask'),
-            "r_input_mask":
-                tf.constant(
-                    all_r_input_masks,
-                    shape=[num_examples, seq_length],
-                    dtype=tf.int32,
-                    name='r_input_mask')}
-        labels_input = None
-        if labels:
-            labels_input = tf.constant(
-                labels,
-                shape=[num_examples],
-                dtype=tf.int32,
-                name='labels')
-        tensor_slices_arg = (feats_dict, labels_input) \
-            if labels_input is not None else feats_dict
-        d = tf.data.Dataset.from_tensor_slices(tensor_slices_arg)
-
-        if is_training:
-            d = d.repeat()
-            d = d.shuffle(buffer_size=100)
-
-        d = d.batch(batch_size=batch_size, drop_remainder=drop_remainder)
-        return d
-
-    return input_fn
-
-
-def input_fn_builder_tfrecords(is_training, drop_remainder, max_seq_length,
-                               tfrecord_save_paths, pad_length=0):
-
-    def input_fn(params):
-        """The actual input function."""
-        batch_size = params["batch_size"]
-
-        # TODO: parameterize this better later on
-
-        # TODO: very hacky to make backwards compatible; consider refactoring
-        raw_dataset = tf.data.TFRecordDataset(tfrecord_save_paths)
-        feature_description = {
-            'l_input_ids': tf.FixedLenFeature([max_seq_length], tf.int64, default_value=None),
-            'r_input_ids': tf.FixedLenFeature([max_seq_length], tf.int64, default_value=None),
-            'l_input_mask': tf.FixedLenFeature([max_seq_length], tf.int64, default_value=None),
-            'r_input_mask': tf.FixedLenFeature([max_seq_length], tf.int64, default_value=None)}
-        if is_training:
-            feature_description['label'] = tf.FixedLenFeature([], tf.int64, default_value=None)
-
-        def _parse(example_proto):
-            parsed_features = tf.parse_single_example(example_proto, feature_description)
-            if is_training:
-                return parsed_features, parsed_features['label']
-            else:
-                return parsed_features
-        d = raw_dataset.map(_parse)
-
-        if is_training:
-            d = d.repeat()
-            d = d.shuffle(buffer_size=100)
-
-        if pad_length and drop_remainder:
-            # hacky padding for batching so nothing actually gets dropped
-            d = d.concatenate(d.take(pad_length))
-
-        d = d.batch(batch_size=batch_size, drop_remainder=drop_remainder)
-        return d
-
-    return input_fn
 
 
 class SiameseBert(object):
@@ -217,76 +109,6 @@ class SiameseBert(object):
         self.random_projection_output_dim = random_projection_output_dim
         self.sum_loss = sum_loss
 
-    def create_model(self, bert_config, is_training, l_input_ids, r_input_ids,
-                     l_input_mask, r_input_mask, labels, use_one_hot_embeddings):
-
-        with tf.variable_scope('bert', reuse=tf.AUTO_REUSE) as bert_scope:
-            l_model = modeling.BertModel(
-                config=bert_config,
-                is_training=is_training,
-                input_ids=l_input_ids,
-                input_mask=l_input_mask,
-                use_one_hot_embeddings=use_one_hot_embeddings,
-                scope=bert_scope,
-                reuse=tf.AUTO_REUSE)
-            r_model = modeling.BertModel(
-                config=bert_config,
-                is_training=is_training,
-                input_ids=r_input_ids,
-                input_mask=r_input_mask,
-                use_one_hot_embeddings=use_one_hot_embeddings,
-                scope=bert_scope,
-                reuse=tf.AUTO_REUSE)
-
-            l_output_layer = l_model.get_pooled_output()
-            r_output_layer = r_model.get_pooled_output()
-
-            with tf.variable_scope('similarity'):
-                l_output_layer = tf.layers.dense(
-                    l_output_layer, self.random_projection_output_dim,
-                    name='random_projection')
-                r_output_layer = tf.layers.dense(
-                    r_output_layer, self.random_projection_output_dim,
-                    name='random_projection', reuse=True)
-
-                l1_norm = -tf.norm(l_output_layer - r_output_layer,
-                                   ord=1,
-                                   axis=-1,
-                                   name='abs_diff')
-                sim_scores = tf.math.exp(l1_norm, name='exp')
-                sim_scores = tf.clip_by_value(sim_scores, 1.0e-7, 1.0-1e-7,
-                                              name='sim_scores')
-                label_preds = tf.math.round(sim_scores, name='label_preds')
-
-            with tf.variable_scope('loss'):
-                logits = tf.math.add(tf.constant(1.0), -sim_scores)
-                logits = tf.math.divide(sim_scores, logits)
-                logits = tf.math.log(logits, name='logits')
-                per_example_loss, loss = None, None
-                if labels is not None:
-                    per_example_loss = tf.nn.sigmoid_cross_entropy_with_logits(
-                        labels=tf.cast(labels, dtype=tf.float32), logits=logits,
-                        name='per_example_loss')
-                    if self.sum_loss:
-                        loss = tf.reduce_sum(per_example_loss, name='total_loss')
-                    else:
-                        loss = tf.reduce_mean(per_example_loss, name='total_loss')
-
-            merged_summaries = None
-
-            if self.feedforward_logging:
-                with tf.name_scope('summaries'):
-                    tf.summary.histogram('sim_scores_hist', sim_scores)
-                    if per_example_loss is not None:
-                        tf.summary.histogram('per_example_loss_hist', per_example_loss)
-                    if loss is not None:
-                        tf.summary.scalar('average_loss', loss)
-
-                    merged_summaries = tf.summary.merge_all(name='network_summaries')
-
-            return (loss, per_example_loss, logits, sim_scores, label_preds,
-                    merged_summaries)
-
     def model_fn_builder(self, bert_config, num_labels, init_checkpoint,
                          learning_rate, use_tpu, use_one_hot_embeddings):
 
@@ -306,9 +128,11 @@ class SiameseBert(object):
             is_training = (mode == tf.estimator.ModeKeys.TRAIN)
 
             (total_loss, per_example_loss, logits, sim_scores, label_preds,
-                merged_summaries) = self.create_model(
+                merged_summaries) = create_model(
                 bert_config, is_training, l_input_ids, r_input_ids,
-                l_input_mask, r_input_mask, labels, use_one_hot_embeddings)
+                l_input_mask, r_input_mask, labels, use_one_hot_embeddings,
+                self.random_projection_output_dim, self.sum_loss,
+                self.feedforward_logging)
             tvars = tf.trainable_variables()
             initialized_variable_names = {}
             scaffold_fn = None
